Reindex works when updating (#958)
Adds a callback to Dashboard::WorksController so that works are
reindexed when they are updated. Note that the callback is not added to
the Work model because this could cause unintended side effect when
reindexing work versions, which also reindex their associated works.

The #destroy endpoint is removed from the controller as well since it is
not used. Works are removed via calls to their work version.