Validate work versions for uniqueness
A work version is identical to another if it has the same metadata and
the same files. We want to prevent users from publishing new versions of
works that don't have any perceptible changes to them. This validation
will happen when publishing, and unless the user has changed something
in the new version, they won't be able to publish it.