Grant ExternalApp admin permissions
ExternalApp behaves very similar to User, and with upcoming changes to
the API, we can use policies to govern what external applications are
allowed to do. For now, they have admin status and can effectively do
anything. But, making that explicit required a few additions to the
class and refactoring/updating the existing policies to ensure that
those permissions we honored.