Find files' S3 etags via Graphql
In order to support checksum verification of files migrated from
Scholarsphere 3, we need to get file metadata information via the API.

This approach uses our graphql endpoint to query for a file using its
NOID from Scholarsphere 3. Because there it not a reliable way to
authorize the file via our Pundit policies, the endpoint query is only
available to administrators.

A user may post a graphql query and request a file's etag from Amazon,
which can be used to verify the file's integrity by calculating an
identical etag from the source in Scholarsphere 3.