Always index asynchronously
Prior to 4.6.0, changes to a work version would redirect back to the
dashboard which was driven by Solr. Due to that, we needed to make sure
that the indexed record for the work version was updated *synchronously*
with the database record so the dashboard would reflect the changes.

Since 4.6.0, work version edits redirect back the resource page, and not
the dashboard. The resource page is backed by the database, so any lag
in Solr indexing really doesn't affect the user. It's possible that if
the user went back to their dashboard immediately after getting
redirected to the resource page, they might not see a newly created or
updated work version; however, that remains to be seen, and the benefits
of a simpler async index process favor future changes such as updating
dois' metadata.