Update Doi metadata
If a work has a doi, its metadata needs to be update when the work is.
This happens when a work, with a doi and a published version, has a new
version published, or an adminitrator updates an existing published
version.

For collections, if the collection has a doi, its metadata will be
updated whenever the collection is.

This also necessitated a change to the DoiService so that only a work's
*published* version would be used for updating its doi.