Remove scholarsphere-client
Changes to the client gem made testing more difficult. Since we were
only using it to test the ingest controller, it makes more sense to take
a different approach and use the internal S3 client to set up the test
instead of the client gem. This was a case where it was causing more
problems than it solved. The integration with the client and
Scholarsphere is still being tested in the client itself.