Always show works and work versions
Draft works and withdrawn works should always resolve. However, their
files are still inaccessible, and draft works are not discoverable when
navigating existing versions in the resource page.

This ensures that any user who has a url for a draft or withdrawn work
will be able to see it, but not discover it via search or browsing other
existing versions.