Allow creation of a new work version from a withdrawn work version (#1043)
In the case where a work only has 1 version and it is withdrawn,
that version is considered the "representative version". The user
can now create a new work version from that withdrawn version.

Previously, if there was only 1 work version and it was withdrawn,
the user would be unable to create any new versions of that work.