Add extracted text derivatives
Adds a derivative file, as defined according to Shrine, to an existing
file resource containing the extracted text found in the original file.
The derivative is created and uploaded via an external applicaiton, in
this case, the metadata-listener, and then the resulting file is added
to the file resource with an API call.