WorkVersionDetailComponent (#1061)
Replaces in the in-view logic with a separate component to determine
messages displayed to the user. If the user can edit the work, and a
draft exists, then the message will give them a link to it.

If the user does not have edit rights, the message only appears if
another published version exists.