Force UTF-8 encoding for extracted text files
Shrine is downloading extracted text derivatives as ASCII-8BIT, but we
need them to be UTF-8 so that they'll index properly. Because
MetadataListener is created the files, we can reasonably assume that
they'll almost always be UTF-8.

Several of the existing file helper classes were refactored and combined
into a new FileHelpers class for better organization in the spec tests.