EditResourceButton component The functionality of the edit and create new work version buttons has been combined into a single button whose action changes depending on whether a draft version exists.
If a draft version exists, the update button acts as "edit draft,"
and if no draft version exists, the update button acts as
"create new draft."

The button to edit a specific work version (draft or otherwise)
is now an admin-only feature.

The Edit button for collections has also been refactored to use
the new EditResourceButton component.