Show messages for draft work versions
If a user is viewing a draft work version and there is a published
version available, show a message directing the user to the latest
published version.

This is an inversion of the previous messaging; we no longer show
a message on published versions that a draft version is available.