background jobs database locking (#1269)
In qa/prod when you upload a thumbnailable? file, the signature job, and the thumbnail job kick off at the same time, and the thumbnail job thomps (this is the technical term) on the file resource. we run both jobs with db table locking so that the thumbnail job will be in queue until the signature job is complete.