Disable the work type selection field in the review & publish step
Since the appearance and behavior of the forms for depositing a work now
depend on the work type, it would be confusing for a user to be able to
change the type, but not see the change in appearance/behavior until
they submit the form. It seems unlikely that a user would need to change
the work type at this stage anyway, but if they did need to for some
reason, they could return to the first step in the workflow where the
work type is initially set.