1089 - focus improvements
Two areas needed to changes in focus to make the Voice Over reader work correctly.
One, when a page is re-rendered due to an error, the error display needs to take focus,
so it is read first. Two, when a new creator or work is added to a collection, it
should focus on the newly created object, where it can be read and edited as needed.