WIP on customizing contributors step of deposit workflow for instruments
This is a work in progress. At this point, I think that the workflow is
working the way that we want it to, but I haven’t written unit tests for
the new form objects yet because I’m not quite happy with the design,
and I was hoping to improve it a bit first. There are also still some
minor to-dos and potential refactorings to address.