Customizing contributors step of deposit workflow for instruments (#1641)
* WIP on customizing contributors step of deposit workflow for instruments

This is a work in progress. At this point, I think that the workflow is
working the way that we want it to, but I haven’t written unit tests for
the new form objects yet because I’m not quite happy with the design,
and I was hoping to improve it a bit first. There are also still some
minor to-dos and potential refactorings to address.

* Refactor some behavior out of the work deposit form base class

The inheritance hierarchy of these form classes is becoming a little
more clear now that there are more of them, and the base class was
collecting too much behavior that wasn’t needed by all of its
subclasses.

* niftany

* Ci fix and fixed pathway merge issue

* Fix some tests

* Rubocop

* Fix broken tests

* Fixed form bugs

* Move contributors form builder into the deposit pathway class

* Adds unit tests for WorkDepositPathway contributor form classes

* Adds header for instruments contributors

* Fix creator building for collections bug

* Refactored/removed .build_form.  .build_creator already has the logic for when there are already creators.  So we just need the contributors form to override it in the instruments pathway.

* Adds controller logic back in to prevent creator build if creators exist

* Update no-op comment

* Delegate creators to workversion even though it wont have any

* add these back in

---------

Co-authored-by: Alex Kiessling <ajkiessl@gmail.com>