Send additional metadata to DataCite for instrument DOIs
I’m not sure what the purpose of the validations in the
DataCite::Metadata classes is. The external DataCite API has its own
validations, as do ScholarSphere’s data models. Without really knowing
the original use case for these validations, the thing that made the
most sense to me was to align new validations with the metadata that we
require for deposit in ScholarSphere and that we want ScholarSphere to
send to DataCite rather than aligning them with the metadata that
DataCite itself requires. Even so, these validations seem like they may
be somewhat redundant, but that could just be coincidental.